refactor: 남은 auth 를 레이어드 아키텍처로 이동
