+++ module-api/src/main/java/kernel/jdon/moduleapi/global/config/auth/JdonLogoutSuccessHandler.java
@@ -10,7 +10,6 @@ import org.springframework.stereotype.Component;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import jakarta.servlet.http.HttpSession;
-import kernel.jdon.auth.util.RedirectUrlUtil;
 import lombok.RequiredArgsConstructor;
 
 @Component
