+++ module-api/src/main/java/kernel/jdon/moduleapi/global/config/web/LoginUserArgumentResolver.java
@@ -1,4 +1,4 @@
-package kernel.jdon.auth.service;
+package kernel.jdon.moduleapi.global.config.web;
 
 import java.util.Objects;
 
@@ -10,8 +10,8 @@ import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.method.support.ModelAndViewContainer;
 
 import jakarta.servlet.http.HttpSession;
+import kernel.jdon.moduleapi.domain.auth.dto.SessionUserInfo;
 import kernel.jdon.moduleapi.global.annotation.LoginUser;
-import kernel.jdon.auth.dto.SessionUserInfo;
 import lombok.RequiredArgsConstructor;
 
 @Component
